lla: move implementation to cpp file
This reduces flash usage
