+++ src/modules/ekf2/EKF/lat_lon_alt/lat_lon_alt.hpp
@@ -73,33 +73,13 @@ public:
 
 	void print() const { printf("latitude = %f (deg), longitude = %f (deg), altitude = %f (m)\n", _latitude_rad, _longitude_rad, (double)_altitude); }
 
-	void operator+=(const matrix::Vector3f &delta_pos)
-	{
-		matrix::Vector2d d_lat_lon_to_d_xy = deltaLatLonToDeltaXY(_latitude_rad, _altitude);
-		_latitude_rad = matrix::wrap_pi(_latitude_rad + static_cast<double>(delta_pos(0)) / d_lat_lon_to_d_xy(0));
-		_longitude_rad = matrix::wrap_pi(_longitude_rad + static_cast<double>(delta_pos(1)) / d_lat_lon_to_d_xy(1));
-		_altitude -= delta_pos(2);
-	}
-
-	void operator+=(const matrix::Vector2f &delta_pos)
-	{
-		matrix::Vector2d d_lat_lon_to_d_xy = deltaLatLonToDeltaXY(_latitude_rad, _altitude);
-		_latitude_rad = matrix::wrap_pi(_latitude_rad + static_cast<double>(delta_pos(0)) / d_lat_lon_to_d_xy(0));
-		_longitude_rad = matrix::wrap_pi(_longitude_rad + static_cast<double>(delta_pos(1)) / d_lat_lon_to_d_xy(1));
-	}
-
-	LatLonAlt operator+(const matrix::Vector3f &delta_pos) const
-	{
-		matrix::Vector2d d_lat_lon_to_d_xy = deltaLatLonToDeltaXY(latitude_rad(), altitude());
-		const double latitude_rad = matrix::wrap_pi(_latitude_rad + static_cast<double>(delta_pos(0)) / d_lat_lon_to_d_xy(0));
-		const double longitude_rad = matrix::wrap_pi(_longitude_rad + static_cast<double>(delta_pos(1)) / d_lat_lon_to_d_xy(1));
-		const float altitude = _altitude - delta_pos(2);
-
-		LatLonAlt lla_new;
-		lla_new.setLatLonRad(latitude_rad, longitude_rad);
-		lla_new.setAltitude(altitude);
-		return lla_new;
-	}
+	/*
+	 * The plus and minus operators below use approximations and should only be used when the Cartesian component is small
+	 */
+	LatLonAlt operator+(const matrix::Vector3f &delta_pos) const;
+	void operator+=(const matrix::Vector3f &delta_pos);
+	void operator+=(const matrix::Vector2f &delta_pos);
+	matrix::Vector3f operator-(const LatLonAlt &lla) const;
 
 	void operator=(const LatLonAlt &lla)
 	{
@@ -108,54 +88,18 @@ public:
 		_altitude = lla.altitude();
 	}
 
-	matrix::Vector3f operator-(const LatLonAlt &lla) const
-	{
-		const double delta_lat = matrix::wrap_pi(_latitude_rad - lla.latitude_rad());
-		const double delta_lon = matrix::wrap_pi(_longitude_rad - lla.longitude_rad());
-		const float delta_alt = _altitude - lla.altitude();
-
-		const matrix::Vector2d d_lat_lon_to_d_xy = deltaLatLonToDeltaXY(_latitude_rad, _altitude);
-		return matrix::Vector3f(static_cast<float>(delta_lat * d_lat_lon_to_d_xy(0)),
-					static_cast<float>(delta_lon * d_lat_lon_to_d_xy(1)),
-					-delta_alt);
-	}
-
 	/*
 	 * Compute the angular rate of the local navigation frame at the current latitude and height
 	 * with respect to an inertial frame and resolved in the navigation frame
 	 */
-	matrix::Vector3f computeAngularRateNavFrame(const matrix::Vector3f &v_ned)
-	{
-		double r_n;
-		double r_e;
-		computeRadiiOfCurvature(_latitude_rad, r_n, r_e);
-		return matrix::Vector3f(
-			       v_ned(1) / (static_cast<float>(r_e) + _altitude),
-			       -v_ned(0) / (static_cast<float>(r_n) + _altitude),
-			       -v_ned(1) * tanf(_latitude_rad) / (static_cast<float>(r_e) + _altitude));
-	}
+	matrix::Vector3f computeAngularRateNavFrame(const matrix::Vector3f &v_ned) const;
 
 private:
 	// Convert between curvilinear and cartesian errors
-	static matrix::Vector2d deltaLatLonToDeltaXY(const double latitude, const float altitude)
-	{
-		double r_n;
-		double r_e;
-		computeRadiiOfCurvature(latitude, r_n, r_e);
-		const double dn_dlat = r_n + static_cast<double>(altitude);
-		const double de_dlon = (r_e + static_cast<double>(altitude)) * cos(latitude);
-
-		return matrix::Vector2d(dn_dlat, de_dlon);
-	}
+	static matrix::Vector2d deltaLatLonToDeltaXY(const double latitude, const float altitude);
 
 	static void computeRadiiOfCurvature(const double latitude, double &meridian_radius_of_curvature,
-					    double &transverse_radius_of_curvature)
-	{
-		const double tmp = 1.0 - pow(Wgs84::eccentricity * sin(latitude), 2);
-		const double sqrt_tmp = std::sqrt(tmp);
-		meridian_radius_of_curvature = Wgs84::meridian_radius_of_curvature_numerator / (tmp * tmp * sqrt_tmp);
-		transverse_radius_of_curvature = Wgs84::equatorial_radius / sqrt_tmp;
-	}
+					    double &transverse_radius_of_curvature);
 
 	struct Wgs84 {
 		static constexpr double equatorial_radius = 6378137.0;
